// samples/studio/src/app/enums/PredefinedExtensionType.ts
export enum PredefinedExtensionType {
  EXTENSION_CLIENT_APP = 'extension-client-app',

  EXTENSION_CLIENT_TAB = 'extension-client-tab',

  EXTENSION_TILE = 'extension-tile',

  EXTENSION_COMPOSER = 'extension-composer',

  EXTENSION_EMAIL = 'extension-email',

  EXTENSION_OAUTH = 'extension-oauth',

  EXTENSION_DATA = 'extension-data',

  EXTENSION_VOICE = 'extension-voice',
}